api/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
asyncpg==0.30.0
pydantic==2.9.2
email-validator==2.1.0.post1
python-dotenv==1.0.1
python-slugify==8.0.4
